+++ src/Console/Command.php
@@ -6,11 +6,9 @@ namespace Pest\Console;
 
 use Pest\Actions\AddsDefaults;
 use Pest\Actions\AddsTests;
+use Pest\Actions\InteractsWithPlugins;
 use Pest\Actions\LoadStructure;
 use Pest\Actions\ValidatesConfiguration;
-use Pest\Contracts\Plugins\AddsOutput;
-use Pest\Contracts\Plugins\HandlesArguments;
-use Pest\Plugin\Loader;
 use Pest\Plugins\Version;
 use Pest\Support\Container;
 use Pest\TestSuite;
@@ -57,23 +55,12 @@ final class Command extends BaseCommand
      */
     protected function handleArguments(array $argv): void
     {
-        /*
-         * First, let's call all plugins that want to handle arguments
-         */
-        $plugins = Loader::getPlugins(HandlesArguments::class);
+        $argv = InteractsWithPlugins::handleArguments($argv);
 
-        /** @var HandlesArguments $plugin */
-        foreach ($plugins as $plugin) {
-            $argv = $plugin->handleArguments($argv);
-        }
-
-        /*
-         * Next, as usual, let's send the console arguments to PHPUnit.
-         */
         parent::handleArguments($argv);
 
         /*
-         * Finally, let's validate the configuration. Making
+         * Let's validate the configuration. Making
          * sure all options are yet supported by Pest.
          */
         ValidatesConfiguration::in($this->arguments);
@@ -128,16 +115,7 @@ final class Command extends BaseCommand
         LoadStructure::in($this->testSuite->rootPath);
 
         $result = parent::run($argv, false);
-
-        /*
-         * Let's call all plugins that want to add output after test execution
-         */
-        $plugins = Loader::getPlugins(AddsOutput::class);
-
-        /** @var AddsOutput $plugin */
-        foreach ($plugins as $plugin) {
-            $result = $plugin->addOutput($result);
-        }
+        $result = InteractsWithPlugins::addOutput($result);
 
         exit($result);
     }