+++ src/Repositories/TestRepository.php
@@ -22,7 +22,7 @@ use PHPUnit\Framework\TestCase;
 final class TestRepository
 {
     /**
-     * @var string
+     * @var non-empty-string
      */
     private const SEPARATOR = '>>>';
 
@@ -42,6 +42,20 @@ final class TestRepository
     public function count(): int
     {
         return count($this->state);
+    }
+
+    /**
+     * Returns the filename of each test that should be executed in the suite.
+     *
+     * @return array<int, string>
+     */
+    public function getFilenames(): array
+    {
+        $testsWithOnly = $this->testsUsingOnly();
+
+        return array_values(array_map(function (TestCaseFactory $factory): string {
+            return $factory->filename;
+        }, count($testsWithOnly) > 0 ? $testsWithOnly : $this->state));
     }
 
     /**
@@ -85,9 +99,7 @@ final class TestRepository
             }
         }
 
-        $onlyState = array_filter($this->state, function ($testFactory): bool {
-            return $testFactory->only;
-        });
+        $onlyState = $this->testsUsingOnly();
 
         $state = count($onlyState) > 0 ? $onlyState : $this->state;
 
@@ -98,6 +110,18 @@ final class TestRepository
                 $each($test);
             }
         }
+    }
+
+    /**
+     * Return all tests that have called the only method.
+     *
+     * @return array<TestCaseFactory>
+     */
+    private function testsUsingOnly(): array
+    {
+        return array_filter($this->state, function ($testFactory): bool {
+            return $testFactory->only;
+        });
     }
 
     /**