+++ tests/PHPUnit/Pest.php
@@ -1,3 +1,0 @@
-<?php
-
-uses(Tests\SubFolder\SubFolder\SubFolder\CustomTestCaseInSubFolder::class)->in('CustomTestCaseInSubFolders/SubFolder');