+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v1.18.0 (2021-08-30)](https://github.com/pestphp/pest/compare/v1.17.0...v1.18.0)
+### Added
+- `toHaveLength` expectation ([#386](https://github.com/pestphp/pest/pull/386))
+- `nunomaduro/collision:^6.0` support ([4ae482c](https://github.com/pestphp/pest/commit/4ae482c7073fb77782b8a4b5738ef1fcea0f82ab))
+
+## [v1.17.0 (2021-08-26)](https://github.com/pestphp/pest/compare/v1.16.0...v1.17.0)
+### Added
+- `toThrow` expectation ([#361](https://github.com/pestphp/pest/pull/361))
+
+## [v1.16.0 (2021-08-19)](https://github.com/pestphp/pest/compare/v1.15.0...v1.16.0)
+### Added
+- Support for new parallel options ([#369](https://github.com/pestphp/pest/pull/369))
+
 ## [v1.15.0 (2021-08-04)](https://github.com/pestphp/pest/compare/v1.14.0...v1.15.0)
 ### Added
 - `toBeTruthy` and `toBeFalsy` ([#367](https://github.com/pestphp/pest/pull/367))